requirements.txt
beautifulsoup4==4.6.3
Click==7.0
Flask==1.0.2
get==1.0.3
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
post==1.0.2
public==1.0.3
python-dotenv==0.9.1
query-string==1.0.2
request==1.0.2
Werkzeug==0.14.1
